Place Saver - save your places on an interactive map
Place Saver was developed by me in December 2013 as a part of a school
course project.
It uses google maps API and allows the user to save different icons with
a short description on the map.
The saved places are stored in the browser's local storage, with the
option to log in to metropolia sql server (students only for now) to
save ur places there.
Possible future extension could be allowing users to make accounts so
they get to keep their saved places across devices.